requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
azure-ai-contentsafety==1.0.0
    # via -r requirements.in
azure-ai-language-conversations==1.1.0
    # via -r requirements.in
azure-ai-textanalytics==5.3.0
    # via -r requirements.in
azure-ai-vision-imageanalysis==1.0.0
    # via -r requirements.in
azure-cognitiveservices-vision-computervision==0.9.1
    # via -r requirements.in
azure-cognitiveservices-vision-customvision==3.1.1
    # via -r requirements.in
azure-common==1.1.28
    # via
    #   azure-ai-textanalytics
    #   azure-cognitiveservices-vision-computervision
    #   azure-cognitiveservices-vision-customvision
azure-core==1.34.0
    # via
    #   azure-ai-contentsafety
    #   azure-ai-language-conversations
    #   azure-ai-textanalytics
    #   azure-ai-vision-imageanalysis
    #   azure-mgmt-core
    #   msrest
azure-mgmt-core==1.5.0
    # via azure-cognitiveservices-vision-customvision
certifi==2025.6.15
    # via
    #   msrest
    #   requests
charset-normalizer==3.4.2
    # via requests
idna==3.10
    # via requests
isodate==0.7.2
    # via
    #   azure-ai-contentsafety
    #   azure-ai-language-conversations
    #   azure-ai-textanalytics
    #   azure-ai-vision-imageanalysis
    #   msrest
msrest==0.7.1
    # via
    #   azure-cognitiveservices-vision-computervision
    #   azure-cognitiveservices-vision-customvision
oauthlib==3.3.0
    # via requests-oauthlib
python-dotenv==1.1.0
    # via -r requirements.in
requests==2.32.4
    # via
    #   azure-core
    #   msrest
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via msrest
six==1.17.0
    # via azure-core
typing-extensions==4.14.0
    # via
    #   azure-ai-language-conversations
    #   azure-ai-textanalytics
    #   azure-ai-vision-imageanalysis
    #   azure-core
urllib3==2.4.0
    # via requests
